(maint) Upload to codecov only when token is present
There is less support of running tokenless codecov upload in the latest
version of the Codecov uploader, and the only places that supports this
quite ofter results in a Rate Limit being hit.

As such, when a token is not present, let us just attempt to do a DryRun
instead of uploading the coverage results.